--- Sources/FrameTagReader.swift
@@ -1,3 +1,0 @@
-import SwiftUI
-import UIKit
-
