+++ Sources/SwiftUI/FrameTag.swift
@@ -1,5 +1,5 @@
 //
-//  PopoverFrameTag.swift
+//  FrameTag.swift
 //  Popovers
 //
 //  Created by A. Zheng (github.com/aheze) on 12/23/21.
@@ -8,13 +8,9 @@
 
 import SwiftUI
 
-/// The key used for the tag-to-frame dictionary.
-struct FrameTag: Hashable {
-    
-    /// The name of the frame.
-    var tag: String
-    
-}
+/**
+ Frame tags are used to store the frames
+ */
 
 /// Store a view's frame for later use.
 struct FrameTagModifier: ViewModifier {
@@ -30,7 +26,6 @@ struct FrameTagModifier: ViewModifier {
                 }
         }
     }
-    
 }
 
 public extension View {
@@ -49,26 +44,27 @@ public extension View {
 }
 
 
-extension UIWindow {
+extension UIResponder {
     
     /**
      Get the saved frame of a frame-tagged view. You must first set the frame using `.frameTag(_:)`.
      - parameter tag: The tag that you used for the frame.
-     
      - Returns: The frame of a frame-tagged view, or `nil` if no view with the tag exists.
      */
     public func frameTagged(_ tag: String) -> CGRect {
         return popoverModel.frameTagged(tag)
     }
     
-    /// Remove all saved frames inside this window for `.popover(selection:tag:attributes:view:)`. Call this method when you present another view where the frames don't apply.
+    /**
+     Remove all saved frames inside this window for `.popover(selection:tag:attributes:view:)`.
+     Call this method when you present another view where the frames don't apply.
+     */
     public func clearSavedFrames() {
         popoverModel.selectionFrameTags.removeAll()
     }
     
     /// Save a frame in this window's `frameTags`.
     func save(_ frame: CGRect, for tag: String) {
-        let frameTag = FrameTag(tag: tag)
-        popoverModel.frameTags[frameTag] = frame
+        popoverModel.frameTags[tag] = frame
     }
 }
