--- Sources/UIKit Conveniences/UIViewController+PresentPopover.swift
@@ -1,11 +1,0 @@
-import UIKit
-
-//extension UIViewController {
-//    
-//    /// Presents the `Popover` using this `UIViewController` as its presentation context.
-//    public func present(_ popoverToPresent: Popover) {
-//        guard let window = view.window else { return }
-//        popoverToPresent.present(in: window)
-//    }
-//    
-//}
